// server/src/types/userInterface.ts
export interface userSelectionsInterface {
  profile?: boolean;
  posts?: boolean;
  comments?: boolean;
  friends?: boolean;
  friends_posts?: boolean;
  friendrequests_to?: boolean;
}

export interface profileSelectionsInterface {
  user?: boolean;
  posts?: boolean;
  comments?: boolean;
  friends?: boolean;
  friends_posts?: boolean;
  friendrequests_to?: boolean;
}